server/requirements.txt
youtube-transcript-api
flask
flask-cors
google-generativeai
google-api-python-client